api/requirements.txt
appdirs==1.4.4
attrs==22.2.0
autopage==0.5.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cliff==4.2.0
cmd2==2.4.3
cryptography==39.0.2
debtcollector==2.5.0
decorator==5.1.1
dogpile.cache==1.1.8
idna==3.4
importlib-metadata==6.0.0
iso8601==1.1.0
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
keystoneauth1==5.1.2
msgpack==1.0.5
netaddr==0.8.0
netifaces==0.11.0
openstacksdk==1.0.1
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==2.7.0
oslo.config==9.1.1
oslo.context==5.1.1
oslo.i18n==6.0.0
oslo.log==5.1.0
oslo.serialization==5.1.1
oslo.utils==6.1.0
packaging==23.0
pbr==5.11.1
prettytable==3.6.0
pycparser==2.21
pyinotify==0.9.6
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.19.3
python-cinderclient==9.3.0
python-dateutil==2.8.2
python-glanceclient==4.3.0
python-keystoneclient==5.1.0
python-neutronclient==9.0.0
python-novaclient==18.3.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requestsexceptions==1.4.0
rfc3986==2.0.0
simplejson==3.18.3
six==1.16.0
stevedore==5.0.0
urllib3==1.26.15
warlock==2.0.1
wcwidth==0.2.6
wrapt==1.15.0
zipp==3.15.0
